Rename Collapsible to GenericCollapsible
