+++ src/components/ui/Filters/Filters.stories.js
@@ -4,7 +4,7 @@ import { storiesOf } from '@storybook/react'
 import State from '../../../utils/State'
 
 import Filters from '.'
-import Collapsible from '../Collapsible'
+import GenericCollapsible from '../GenericCollapsible'
 import GenericTabs from '../GenericTabs'
 import InputField from '../../forms/InputField'
 import SelectBox from '../SelectBox'
@@ -21,11 +21,11 @@ storiesOf('UI Components/Filters', module)
   .add('default display', () => (
     <Filters.Container>
 
-      <Collapsible.Container initialExpanded>
-        <Collapsible.Header>
+      <GenericCollapsible.Container initialExpanded>
+        <GenericCollapsible.Header>
           <Filters.Heading>Context Filters</Filters.Heading>
-        </Collapsible.Header>
-        <Collapsible.Body>
+        </GenericCollapsible.Header>
+        <GenericCollapsible.Body>
           <Filters.Group>
             <Filters.Label>Show</Filters.Label>
             <GenericTabs.Container defaultActiveKey='list' headWrapper={Filters.TabHeads}>
@@ -70,14 +70,14 @@ storiesOf('UI Components/Filters', module)
               </GenericTabs.Pane>
             </GenericTabs.Container>
           </Filters.Group>
-        </Collapsible.Body>
-      </Collapsible.Container>
+        </GenericCollapsible.Body>
+      </GenericCollapsible.Container>
 
-      <Collapsible.Container initialExpanded>
-        <Collapsible.Header>
+      <GenericCollapsible.Container initialExpanded>
+        <GenericCollapsible.Header>
           <Filters.Heading>Search Filters</Filters.Heading>
-        </Collapsible.Header>
-        <Collapsible.Body>
+        </GenericCollapsible.Header>
+        <GenericCollapsible.Body>
           <Filters.Group>
             <Filters.Filter label='Status'>
               <SelectBoxFilterContainer placeholder='Status' />
@@ -92,8 +92,8 @@ storiesOf('UI Components/Filters', module)
               <InputFilterContainer type='number' />
             </Filters.Filter>
           </Filters.Group>
-        </Collapsible.Body>
-      </Collapsible.Container>
+        </GenericCollapsible.Body>
+      </GenericCollapsible.Container>
 
     </Filters.Container>
   ))
